moved some common scripts outside of year specific scripts
